refactor: move the namespace serializers at the root
