+++ Expressif.Testing/ExpressionBuilderTest.cs
@@ -1,7 +1,6 @@
-﻿using Expressif.Functions;
-using Expressif.Functions.Serializer;
-using Expressif.Functions.Text;
+﻿using Expressif.Functions.Text;
 using Expressif.Parsers;
+using Expressif.Serializers;
 using Expressif.Values.Resolvers;
 using System;
 using System.Collections.Generic;
