+++ Expressif.Testing/PredicationBuilderTest.cs
@@ -1,8 +1,6 @@
-﻿using Expressif.Functions.Text;
-using Expressif.Parsers;
-using Expressif.Predicates.Operators;
-using Expressif.Predicates.Serializer;
+﻿using Expressif.Parsers;
 using Expressif.Predicates.Text;
+using Expressif.Serializers;
 using NUnit.Framework;
 using System;
 using System.Collections.Generic;
