+++ Expressif/ExpressionBuilder.cs
@@ -1,15 +1,7 @@
 ﻿
-/* Unmerged change from project 'Expressif (net6.0)'
-Before:
-using Expressif.Parsers;
-After:
-using Expressif;
-using Expressif;
-using Expressif.Parsers;
-*/
 using Expressif.Functions;
-using Expressif.Functions.Serializer;
 using Expressif.Parsers;
+using Expressif.Serializers;
 using Expressif.Values.Special;
 using System;
 using System.Collections.Generic;
@@ -17,7 +9,6 @@ using System.Linq;
 using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
-using static System.Net.Mime.MediaTypeNames;
 
 namespace Expressif
 {
