+++ Expressif/Serializers/ParameterSerializer.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Expressif.Functions.Serializer
+namespace Expressif.Serializers
 {
     public class ParameterSerializer
     {
